+++ be/ebook-search/ebook-web/src/docs/asciidoc/admin-book.adoc
@@ -1,0 +1,13 @@
+= 관리자 도서 API
+
+== 책 저장 API
+
+operation::create-book[snippets='http-request,request-body,http-response,response-headers']
+
+== 책 업데이트 API
+
+operation::update-book[snippets='http-request,path-parameters,request-body,http-response,response-headers']
+
+== 책 삭제 API
+
+operation::delete-book[snippets='http-request,path-parameters,http-response']
