+++ be/ebook-search/ebook-web/src/docs/asciidoc/index.adoc
@@ -2,8 +2,14 @@
 
 == 목차
 
+=== User API
 * link:book.html[도서 API]
 * link:library.html[도서관 API]
+
+=== Admin API
+* link:admin-book.adoc[관리자 도서 API]
+* link:admin-library.adoc[관리자 도서관 API]
+* link:admin-verification.adoc[검증 API]
 
 == 개요
