+++ be/ebook-search/ebook-web/src/main/java/com/meetyourbook/controller/admin/AdminBookController.java
@@ -1,0 +1,56 @@
+package com.meetyourbook.controller.admin;
+
+import com.meetyourbook.dto.BookCreateRequest;
+import com.meetyourbook.dto.BookUpdateRequest;
+import com.meetyourbook.service.BookService;
+import jakarta.validation.Valid;
+import java.net.URI;
+import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
+@RestController
+@RequestMapping("/admin/books")
+@RequiredArgsConstructor
+public class AdminBookController {
+
+    private final BookService bookService;
+
+    @PostMapping
+    public ResponseEntity<Void> createBook(@RequestBody @Valid BookCreateRequest request) {
+        UUID id = bookService.createBook(request);
+        URI location = ServletUriComponentsBuilder
+            .fromCurrentRequest()
+            .path("/{id}")
+            .buildAndExpand(id)
+            .toUri();
+        return ResponseEntity.created(location).build();
+    }
+
+    @DeleteMapping("/{bookId}")
+    public ResponseEntity<Void> deleteBook(@PathVariable UUID bookId) {
+        bookService.deleteBook(bookId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @PatchMapping("/{bookId}")
+    public ResponseEntity<?> updateBook(@PathVariable UUID bookId,
+        @RequestBody @Valid BookUpdateRequest request) {
+        bookService.updateBook(bookId, request);
+        URI location = ServletUriComponentsBuilder
+            .fromCurrentRequest()
+            .path("/{id}")
+            .buildAndExpand(bookId)
+            .toUri();
+        return ResponseEntity.noContent().location(location).build();
+    }
+
+}
